+++ tests/testthat/test-translate-odbc.R
@@ -21,9 +21,6 @@ test_that("custom aggregators translated correctly", {
   expect_equal(trans(sd(x)),         sql("STDDEV_SAMP(`x`)"))
   expect_equal(trans(count()),       sql("COUNT(*)"))
   expect_equal(trans(n()),           sql("COUNT(*)"))
-  expect_equal(trans(n_distinct(x)), sql("COUNT(DISTINCT `x`)"))
-
-
 })
 
 test_that("custom window functions translated correctly", {
@@ -35,9 +32,6 @@ test_that("custom window functions translated correctly", {
   expect_equal(trans(sd(x, na.rm = TRUE)),         sql("STDDEV_SAMP(`x`) OVER ()"))
   expect_equal(trans(count()),       sql("COUNT(*) OVER ()"))
   expect_equal(trans(n()),           sql("COUNT(*) OVER ()"))
-  expect_equal(trans(n_distinct(x)), sql("COUNT(DISTINCT `x`) OVER ()"))
-
-
 })
 
 test_that("queries translate correctly", {